+++ tests/faces/test_planeface3d.py
@@ -46,7 +46,6 @@ class TestPlaneFace3D(unittest.TestCase):
         R = 0.15
         cylindricalsurface = volmdlr.faces.CylindricalSurface3D(volmdlr.OXYZ, R)
         face = cylindricalsurface.rectangular_cut(0, volmdlr.TWO_PI, -.25, .25)
-        plane = volmdlr.faces.Plane3D(volmdlr.OZXY)
         """ ========== CIRCLE3D ========="""
         plane_face_3 = self.plane_face_cylindricalface_intersec.rotation(volmdlr.O3D, volmdlr.X3D, math.pi / 2)
         face_intersections = plane_face_3.face_intersections(face)
@@ -61,8 +60,8 @@ class TestPlaneFace3D(unittest.TestCase):
         self.assertIsInstance(face_intersections[0].primitives[0], wires.Ellipse3D)
         self.assertEqual(face_intersections[0].primitives[0].center, volmdlr.O3D)
         self.assertAlmostEqual(face_intersections[0].primitives[0].major_axis, 0.21213203435596426)
-        self.assertEqual(face_intersections[0].primitives[0].major_dir,
-                         volmdlr.Vector3D(0, 0.7071067811865475, -0.7071067811865475))
+        self.assertTrue(face_intersections[0].primitives[0].major_dir.is_close(volmdlr.Vector3D(0, 0.7071067811865475,
+                                                                                                -0.7071067811865475)))
         """ ========== THREE ARC ELLIPSES ========="""
         plane_face_3 = self.plane_face_cylindricalface_intersec.rotation(volmdlr.O3D, volmdlr.X3D, math.pi / 7)
         face_intersections = plane_face_3.face_intersections(face)
@@ -82,9 +81,9 @@ class TestPlaneFace3D(unittest.TestCase):
                                  volmdlr.Point3D(-0.08947272158306664, 0.12039365470206077, -0.25)]]
         for expected_points, wire in zip(list_expected_points, face_intersections):
             arcellipse = wire.primitives[0]
-            self.assertEqual(expected_points[0], arcellipse.start)
-            self.assertEqual(expected_points[1], arcellipse.interior)
-            self.assertEqual(expected_points[2], arcellipse.end)
+            self.assertTrue(expected_points[0].is_close(arcellipse.start))
+            self.assertTrue(expected_points[1].is_close(arcellipse.interior))
+            self.assertTrue(expected_points[2].is_close(arcellipse.end))
         """ ========== TWO PARALLEL LINES ========="""
         plane_face_3 = self.plane_face_cylindricalface_intersec.rotation(volmdlr.O3D, volmdlr.X3D, math.pi)
         face_intersections = plane_face_3.face_intersections(face)