+++ CHANGELOG.md
@@ -13,11 +13,19 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 
+
 ### Removed
 
 ### Performance improvements
+* eq & hash: Some eq and hash methods have been fixed. starting from clases Point and Vector.
 
 ### Refactorings
+
+### Unittests
+* Unittests for Vector2D
+* Unittests for Point2D
+* Unittests for Vector3D
+* Unittests for Point3D
 
 
 ## v0.9.0 [Testing]